+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## v0.1.0
+
+### Added
+- Added function to get storage information
+- Added function to get count of HA nodes
+- Split functions into seperate files
+
+## v0.0.2
+
+### Updated
+- Added error handling to functions
+
 ## v0.0.1
 
 ### Added